requirements.txt
arrow==0.11.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
docxtpl==0.6.3
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.14.0
ics==0.5
idna==2.8
Jinja2==2.10.3
lxml==4.4.1
MarkupSafe==1.1.1
numpy==1.17.2
oauthlib==3.1.0
pandas==0.25.1
pyasn1==0.4.7
pyasn1-modules==0.2.7
python-dateutil==2.8.0
python-docx==0.8.7
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.25.6
